requirements.txt
fastapi
uvicorn
python-dotenv
motor
pymongo
python-multipart
requests
pydantic

numpy
opencv-python
deepface
tensorflow
torch
torchvision